chore: localize serde identity helper (#865)
