+++ crates/primitives/Cargo.toml
@@ -10,7 +10,6 @@ license.workspace = true
 bs58.workspace = true
 borsh = { workspace = true, features = ["derive"], optional = true }
 ed25519-dalek.workspace = true
-libp2p-identity = { workspace = true, features = ["peerid", "serde"] }
 rand = { workspace = true, optional = true }
 serde = { workspace = true, features = ["derive"] }
 serde_json.workspace = true
